Add chapter name to history view
Capturing the chapter name in the history view is helpful because it can
greatly increase the time it takes for the user to grok the location
since it is usually more meaningful context indicator than timestamp
alone.

This commit adds the chapter name to the history, displayed just to the
left of the timestamp.  If the chapter name is too limit, it is
truncated with ellipsis (...)

Fixes #1766